refactor: 뷰포트 사이즈 체크 커스텀 훅 수정
